requirements.txt
fastapi
fastapi-mcp
uvicorn
mcp-proxy
python-dotenv
httpx